Improve report generation feature.
