+++ components/dashboards-web-component/src/viewer/DashboardViewPage.jsx
@@ -47,6 +47,7 @@ import html2cavas from 'html2canvas';
 import axios from 'axios';
 import { pdfConfig } from './components/JspdfConf.js';
 import AuthManager from '../auth/utils/AuthManager';
+import ReportGeneration from '../utils/ReportGeneration';
 
 const SelectableList = makeSelectable(List);
 
@@ -90,12 +91,12 @@ class DashboardViewPage extends Component {
 
         this.handleChange = this.handleChange.bind(this);
         this.capturePage = this.capturePage.bind(this);
-        this.generatePdf = this.generatePdf.bind(this);
         this.handleCardClick = this.handleCardClick.bind(this);
         this.handleClose = this.handleClose.bind(this);
         this.handleOpen = this.handleOpen.bind(this);
-        this.handleincludeGenerateTime = this.handleincludeGenerateTime.bind(this);
+        this.handleIncludeGenerateTime = this.handleIncludeGenerateTime.bind(this);
         this.removePage = this.removePage.bind(this);
+        this.generateDashboardReport=this.generateDashboardReport.bind(this);
     }
 
     componentDidMount() {
@@ -178,7 +179,7 @@ class DashboardViewPage extends Component {
             <FlatButton
                 label='Export'
                 primary={true}
-                onClick={this.generatePdf}
+                onClick={this.generateDashboardReport}
             />,
         ];
 
@@ -214,7 +215,7 @@ class DashboardViewPage extends Component {
                 >
                     <Checkbox
                         label='Include report genetation time'
-                        onClick={this.handleincludeGenerateTime}
+                        onClick={this.handleIncludeGenerateTime}
                     />
                 </Dialog>
 
@@ -434,7 +435,7 @@ class DashboardViewPage extends Component {
         this.setState({ dialogOpen: true });
     }
 
-    handleincludeGenerateTime() {
+    handleIncludeGenerateTime() {
         this.setState({ includeTime: !this.state.includeTime });
     }
 
@@ -448,178 +449,31 @@ class DashboardViewPage extends Component {
 
     progress = () => {
         const { completed } = this.state;
-        this.setState({ completed: completed >= 100 ? 0 : completed + 10 });
+        this.setState({ completed: completed >= 100 ? 0 : completed + 10});
     };
-
-    formatDate(date) {
-        let hours = date.getHours();
-        let minutes = date.getMinutes();
-        const ampm = hours >= 12 ? 'PM' : 'AM';
-        let day = date.getDate();
-        let month = date.getMonth() + 1;
-        hours = hours % 12;
-        hours = hours ? hours : 12; // the hour '0' should be '12'
-        minutes = minutes < 10 ? '0' + minutes : minutes;
-        day = day < 10 ? '0' + day : day;
-        month = month < 10 ? '0' + month : month;
-        const timeStr = hours + ':' + minutes + ' ' + ampm;
-        return day + '/' + month + '/' + date.getFullYear() + '  ' + timeStr;
-    }
 
     sleep(ms) {
         return new Promise(resolve => setTimeout(resolve, ms));
     }
 
-    async generatePdf() {
+    async generateDashboardReport(){
 
         //Starts showing report generation progress
         this.handleClose();
         this.handleReportStatusOpen();
-        this.timer = setInterval(this.progress, 195);
-
-        const pdf = new jspdf('l', 'px', this.state.pageSize.toLowerCase());
+        this.timer = setInterval(this.progress, 192);
+        await this.sleep(2000);
 
         const title = document.getElementsByTagName('h1')[0].innerText;
 
-        // Get header : a cutomized header defined by the user or else the default stream processor header
-        await this.addPdfImage(`/${appContext}/apis/dashboards/pdfHeader`, function (res) {
-            if (res != 'none') {
-                localStorage.setItem('dashboardHeader', res);
-            }
-        });
-
-        // Get footer : a cutomized footer defined by the user or else the no footer by default
-        await this.addPdfImage(`/${appContext}/apis/dashboards/pdfFooter`, function (res) {
-            if (res != 'none') {
-                localStorage.setItem('dashboardFooter', res);
-            }
-        });
-
-        // To handle the async calls (wait until the image data is stored in localStorage)
-        await this.sleep(2000);
-
-        const img = localStorage.getItem('dashboardHeader');
-        pdf.addImage(img, 'JPEG', pdfConfig.stampImageDashboard.coordinates.x,
-                     pdfConfig.stampImageDashboard.coordinates.y, pdfConfig.stampImageDashboard.size.x,
-                     pdfConfig.stampImageDashboard.size.y);
-
-        this.state.pageList.map((item, ind) => {
-            const it = localStorage.getItem(item);
-            const image = JSON.parse(it);
-            const imgData = image.imageData;
-
-            const imgWidth = image.width;
-            const imgHeight = image.height;
-
-            const page = this.dashboard.pages.find(page => {
-                return page.id == item;
-            })
-
-            //Add title
-            pdf.setFontSize(15);
-            pdf.setFontType('bold');
-            const padding = Array.apply(null, Array(title.length + 4)).map(function () { return '  ' })
-            pdf.text((pdf.internal.pageSize.getHeight() / 2 + 30), 18, title + ' : ');
-
-            const ntext = padding.join('') + page.name;
-            pdf.setFontType('normal');
-            pdf.text((pdf.internal.pageSize.getHeight() / 2 + 30), 18, ntext);
-
-            let pdfInfo = '';
-
-            if (this.state.includeTime) {
-                pdfInfo = 'Generated on : ' + this.formatDate(new Date());
-            }
-
-            pdf.setFontType('bold');
-            pdf.setFontSize(12);
-            pdf.text(pdfInfo, pdfConfig.stampImageDashboard.coordinates.x, 35);
-            pdf.setFontType('normal');
-
-            if (imgWidth > imgHeight) {
-                const val = (pdf.internal.pageSize.getWidth()) / imgWidth;
-                if (imgHeight * val < (pdf.internal.pageSize.getHeight()-60)) {
-                    pdf.addImage(imgData, 'JPEG', 0, 55, imgWidth * val, imgHeight * val, 'dashboard' + ind, 'FAST');
-                } else {
-                    const valH = (pdf.internal.pageSize.getHeight()-60) / imgHeight;
-                    const xposition = (pdf.internal.pageSize.getWidth() - imgWidth * valH) / 2;
-                    pdf.addImage(imgData, 'JPEG', xposition, 45, imgWidth * valH, imgHeight * valH, 'dashboard' + ind,
-                    'FAST');
-                }
-            } else {
-                const val = (pdf.internal.pageSize.getHeight()) / imgHeight;
-                pdf == new jspdf('p', 'px', this.state.pageSize.toLowerCase());
-                pdf.addImage(imgData, 'JPEG', 0, 35, imgWidth * val, imgHeight * val, 'dashboard' + ind, 'FAST');
-            }
-
-            // FOOTER : a customized footer defined by the user in the deployment.yaml file or else no footer by default
-            const pageNo = 'Page ' + (ind + 1) + ' of ' + this.state.pageList.length;
-            pdf.setFontSize(10);
-            const pageHeight = pdf.internal.pageSize.height || pdf.internal.pageSize.getHeight();
-            pdf.text(pageNo, 10, pageHeight - 10);
-
-            const footerImg = localStorage.getItem('dashboardFooter');
-            if (footerImg != null) {
-                pdf.addImage(footerImg, 'JPEG', 380, 780, pdfConfig.stampImage.size.x, pdfConfig.stampImage.size.y);
-            }
-
-            if (ind < this.state.pageList.length - 1) {
-                pdf.addPage();
-
-                // HEADER : a cutomized header defined by the user or else the default stream processor header
-                const headerImg = localStorage.getItem('dashboardHeader');
-                if (headerImg != null) {
-                    pdf.addImage(headerImg, 'JPEG', pdfConfig.stampImageDashboard.coordinates.x,
-                                 pdfConfig.stampImageDashboard.coordinates.y, pdfConfig.stampImageDashboard.size.x,
-                                 pdfConfig.stampImageDashboard.size.y);
-                }
-            }
-        });
+        const pdf=ReportGeneration.generatePdf(this.state.pageSize.toLowerCase(),this.state.pageList,
+                                                this.state.includeTime,title,this.dashboard.pages);
 
         clearInterval(this.timer);
         this.handleReportStatusClose();
         this.state.completed = 0;
+        this.state.includeRecords = false;
         this.state.includeTime = false;
-
-        const docName = title + '.pdf';
-        pdf.save(docName);
-
-    }
-
-    // Get the image from deployment.yaml file through the server
-    async addPdfImage(url, callback) {
-        let path = `/${appContext}/public/app/images/`;
-
-        const httpClient = axios.create({
-            baseURL: url,
-            timeout: 2000,
-            headers: { Authorization: 'Bearer ' + AuthManager.getUser().SDID },
-        });
-
-        await httpClient.get()
-            .then(function(res){
-                if (res.data === '') {
-                    callback('none');
-                } else {
-                    path += res.data;
-
-                    //Convert image to bdase64 encoded image (data_url)
-                    const canvas = document.createElement('canvas');
-                    const ctx = canvas.getContext('2d');
-                    const img = new Image();
-                    let imgStr;
-
-                    img.onload = function () {
-                        canvas.width = img.width;
-                        canvas.height = img.height;
-                        ctx.drawImage(img, 0, 0);
-                        imgStr = canvas.toDataURL('image/png');
-                        callback(imgStr);
-                    }
-
-                    img.src = path;
-                }
-            });
 
     }
